BaseTools: Disable Second DLINK step for WHOLEARCHIVE
Project Mu does not use WHOLEARCHIVE, due to issues in silicon
partner code when it is enabled. In order to support WHOLEARCHIVE,
edk2 added a second dlink step in the build rules to get the
benefit of identifying duplicate symbols without needing the
space expansion. This is an extra linking step for each module
and so does increase build time. Without WHOLEARCHIVE, this is
an unneeded step on Mu, so it is disabled here.